Extract component registry from config
