+++ DashAI/back/api/api_v1/endpoints/components.py
@@ -6,7 +6,7 @@ from fastapi import APIRouter, Query, status
 from fastapi.exceptions import HTTPException
 from typing_extensions import Annotated
 
-from DashAI.back.core.config import component_registry
+from DashAI.back.core.component_registry import component_registry
 
 logging.basicConfig(level=logging.DEBUG)
 log = logging.getLogger(__name__)
