+++ tests/back/api/test_components_api.py
@@ -2,7 +2,7 @@
 import pytest
 from fastapi.testclient import TestClient
 
-from DashAI.back.core.config import component_registry
+from DashAI.back.core.component_registry import component_registry
 from DashAI.back.dataloaders import BaseDataLoader
 from DashAI.back.models import BaseModel
 from DashAI.back.registries import ComponentRegistry
