+++ tests/back/api/test_runner_api.py
@@ -5,7 +5,7 @@ import pytest
 from fastapi.testclient import TestClient
 from sqlalchemy.orm import sessionmaker
 
-from DashAI.back.core.config import component_registry
+from DashAI.back.core.component_registry import component_registry
 from DashAI.back.core.runner import RunnerError
 from DashAI.back.database.models import Experiment, Run
 from DashAI.back.dataloaders.classes.csv_dataloader import CSVDataLoader
